pyproject.toml
[project]
name = "nus-fintech-asset-recommender"
version = "0.1.0"
description = ""
authors = [
    {name = "yjiahao",email = "yjiahao123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "pandas (>=2.3.2,<3.0.0)",
    "numpy (==1.26.4)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "gradio (>=5.49.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "supabase (>=2.24.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
